+++ banking-api/src/main/java/io/dabrowa/whitebox/api/events/AccountCreatedEvent.java
@@ -2,7 +2,9 @@ package io.dabrowa.whitebox.api.events;
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
+import java.math.BigDecimal;
+
 public record AccountCreatedEvent(@TargetAggregateIdentifier String accountId,
-                                  long initialBalance,
-                                  long overdraftLimit) implements Event {
+                                  BigDecimal initialBalance,
+                                  BigDecimal overdraftLimit) implements Event {
 }
