+++ banking-api/src/main/java/io/dabrowa/whitebox/api/events/AccountCreditedEvent.java
@@ -2,5 +2,7 @@ package io.dabrowa.whitebox.api.events;
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
-public record AccountCreditedEvent(@TargetAggregateIdentifier String accountNumber, long creditValue, long creditedAtEpochMillis) {
+import java.math.BigDecimal;
+
+public record AccountCreditedEvent(@TargetAggregateIdentifier String accountNumber, BigDecimal creditValue, long creditedAtEpochMillis) {
 }
