+++ banking-api/src/main/java/io/dabrowa/whitebox/api/events/AccountDebitedEvent.java
@@ -2,5 +2,7 @@ package io.dabrowa.whitebox.api.events;
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
-public record AccountDebitedEvent(@TargetAggregateIdentifier String accountNumber, long debitValue, long creditedAtEpochMillis) {
+import java.math.BigDecimal;
+
+public record AccountDebitedEvent(@TargetAggregateIdentifier String accountNumber, BigDecimal debitValue, long creditedAtEpochMillis) {
 }
