+++ banking-api/src/main/java/io/dabrowa/whitebox/api/queries/AccountBalanceQuery.java
@@ -1,6 +1,8 @@
 package io.dabrowa.whitebox.api.queries;
 
+import java.math.BigDecimal;
+
 public record AccountBalanceQuery(String accountId) {
-    public record BalanceQueryResponse(long balance) {
+    public record BalanceQueryResponse(BigDecimal balance) {
     }
 }
